+++ .env.example
@@ -9,3 +9,4 @@ ROOT_LOG_LEVEL=INFO
 SECURITY_API_KEY_SECURED_PATHS="/hello"
 SECURITY_API_KEY_PRIMARY=primary-key
 SECURITY_API_KEY_SECONDARY=secondary-key
+API_REST_END_POINT="https://api.dev.platform.pagopa.it/afm/calculator-service"
