+++ .gitignore
@@ -160,4 +160,4 @@ gradle-app.setting
 # End of https://www.toptal.com/developers/gitignore/api/intellij+all,kotlin,java,gradle
 /.env
 /checkouts/*
-/mock-server/node_modules/
+/mock-server/*
