+++ src/main/java/it/pagopa/ecommerce/GreetingResource.java
@@ -1,0 +1,1 @@
+
