+++ src/main/resources/application.properties
@@ -18,9 +18,10 @@ quarkus.log.console.json.additional-field."service.environment".value=${ECS_SERV
 security.apiKey.primary=${SECURITY_API_KEY_PRIMARY}
 security.apiKey.secondary=${SECURITY_API_KEY_SECONDARY}
 security.apiKey.securedPaths=${SECURITY_API_KEY_SECURED_PATHS}
-quarkus.rest-client.payment-method.url=http://localhost:9090
+quarkus.rest-client.payment-method.url=${API_REST_END_POINT}
 
-quarkus.openapi-generator.codegen.input-base-dir=api-spec
+quarkus.openapi-generator.codegen.input-base-dir=api-client
+
 quarkus.openapi-generator.codegen.spec.api_yaml.mutiny=true
